+++ NEWS.md
@@ -1,6 +1,12 @@
 
 # renv 0.18.0  (UNRELEASED)
 
+* `renv` gains a new function `renv::checkout()`, for installing the
+  latest-available packages from a selected repository. For example, one can
+  use `renv::checkout(date = "2023-02-08")` to install the packages available
+  on 2023-02-08 from the [Posit Public Package Manager](https://packagemanager.rstudio.com/)
+  repository.
+  
 * `renv::install()` now supports Bioconductor remotes of the form
   `bioc::<BiocVersion>/<Package>`, for installing packages from
   a particular version of Bioconductor. Aliases like 'release' and